+++ src/main_test.h
@@ -9,8 +9,8 @@
 #define ALLOCATE_THE_UNIVERSE_CHUNK_SIZE (634 * 1024) // 634K
 //#define ALLOCATE_THE_UNIVERSE_CHUNK_SIZE (1024 * 1024) // 1M
 
-size_t test_dmm_malloc(void);
-size_t test_dmm_free_sets_header(void);
-size_t test_dmm_allocate_the_universe(void);
+int test_dmm_malloc(void);
+int test_dmm_free_sets_header(void);
+int test_dmm_allocate_the_universe(void);
 
 #endif
